match underlying type of Optional to encoding type

+++ common.h
@@ -14,7 +14,7 @@
 
 #include "io.h"
 
-enum class Optional : short {
+enum class Optional : uint8_t {
   undefined = 0
 };
